requirements.txt
torch
transformers
faiss-cpu
pandas
numpy
pillow
requests
tqdm
scikit-learn
